dashboard/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
databricks-sql-connector==3.3.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
